app/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.10
kafka-python==2.0.2
python-json-logger==2.0.4